// src/hooks/types.tsx
export type BalancesType = {
  tcapBalance: string;
  jpegzBalance: string;
  ctxBalance: string;
  tcapSupplly: string;
  jpegzSupplly: string;
};

export type OraclePricesType = {
  jpegzOraclePrice: string;
  jpegzMarketCap: string;
  tcapOraclePrice: string;
  tcapMarketCap: string;
  wethOraclePrice: string;
  daiOraclePrice: string;
  aaveOraclePrice: string;
  linkOraclePrice: string;
  uniOraclePrice: string;
  snxOraclePrice: string;
  maticOraclePrice: string;
  wbtcOraclePrice: string;
  usdcOraclePrice: string;
};